+++ providers/dns/tencentcloud/tencentcloud.toml
@@ -14,8 +14,9 @@ lego --email you@example.com --dns tencentcloud --domains my.example.org run
   [Configuration.Credentials]
     TENCENTCLOUD_SECRET_ID = "Access key ID"
     TENCENTCLOUD_SECRET_KEY = "Access Key secret"
-    TENCENTCLOUD_SESSION_TOKEN = "Access Key token"
   [Configuration.Additional]
+    TENCENTCLOUD_SESSION_TOKEN = "Access Key token"
+    TENCENTCLOUD_REGION = "Region"
     TENCENTCLOUD_POLLING_INTERVAL = "Time between DNS propagation check"
     TENCENTCLOUD_PROPAGATION_TIMEOUT = "Maximum waiting time for DNS propagation"
     TENCENTCLOUD_TTL = "The TTL of the TXT record used for the DNS challenge"
